Fix a huge number of bugs and linter warnings

+++ include/internal/common_util.h
@@ -14,7 +14,7 @@
 
 void copy_int_as_le(void *dst, void *src, size_t len);
 
-int validate_path_component(const char *cmpnt, size_t len_s);
+int validate_path_component(const char *cmpnt, uint8_t len_s);
 
 inline void *offset_ptr(void *ptr, size_t off) {
     return (void*) ((uintptr_t) ptr + off);
